ser: Fix union serialization return type issue
The problem is that even though we unroll each field in the union type
and compare their name to the value passed in, the compiler can't figure
that out for some reason, so it thinks there's a chance we don't call
serialize.

However, we know that can't possibly happen, so we can just put an
unreachable after the loop.