de: Make visitor methods generic, add access types
The first part of this commit renames the `Ok` declaration within
visitors to `Value`. It makes more sense since all visitors produce some
value.

The next part consists of updating the methods required by the `Visitor`
interface to be generic over an error instead of the interface itself.
This allows the visitor to be used by any deserializer, without having
to create a new visitor type for each deserializer.

The final portion of this commit adds the *Access interface types.