Reorg ser
Okay so, different approach to the problem addressed by the last commit.
In this commit, we simply namespace out the impls and interface decls
instead of having the serialize functions in lib.zig. I actually think
it does make sense to have them under ser for aesthetic and relational
reasons.