Use usingnamespace interface impl design
There were two problems with this approach before: 1) how to get the
interface type and 2) how to customize the interface function name.

The second is easily solved by simply namespacing the returned function
and then creating an alias for it.

The first can be solved by using @typeInfo to get the return type
of the returned function.