ser: Remove getty.free references
The free function makes certain assumptions. Namely, it assumes that
pointers of any kind are allocated. This means that if you have a
pointer to a string literal, it probably will be wrongly freed by
'getty.free'.

The function should only be used for freeing resources created by
'getty.deserialize' and 'getty.deserializeWith'.