de: Make allocators unmanaged
The deserialize and visit methods of the Deserializer and Visitor
interface now take allocator parameters. They are no longer managed by
implementations.

The motivation for this change was so that visitor implementations can
just be structs with no fields. That way, Getty would always be able to
create instances of them, which means the visitor function in DBs can
return a visitor type instead of an arbitrary implementation instance,
and Getty would handle creating the instance.