ser: Refactor tests
Tests are now in-file alongside the implementations. Also the build process
has been updated so that you can now run 'zig build test-ser' to test only
serialization-related code. This required removing all references to lib.zig
in ser.zig and ser/, otherwise the de decls that were referenced in lib.zig
and the files it imported would cause the de tests to be ran as well.

Traits have also been moved out of the concepts namespace.