ser: Remove serialization tuple length check
user_st and serializer_st are set to an empty tuple if getty.default_st is used
as their value. This would cause the length check to trigger in cases such as
when an internal serializer (e.g., MapKeySerializer in getty-json) would pass in
the external serializer's user_st to the getty.Serializer interface.