ser: Fixes for type-defined blocks
Type-defined serialization tuples are no longer a thing. Tuples are for
grouping serialization blocks that apply to different types. However, a
type-defined block only ever applies to the type that defined it, so it
makes no sense for a type-defined tuple to exist.

Also, type-defined serialization blocks no longer require an is
function. As mentioned earlier, they always apply to the type that
defined them, so there's no need for such a function.

Separate is_tsb and has_sb functions have been added for type-defined
serialization blocks.