ser: Fix ArrayBitSet serialization
The logic for serializing unset bits after the initial serialization
loop was incorrect in that would never serialize the last bit as 0 if
the bit in position 1 was set.

This was because the 0th bit serialization logic relied on the i
variable, which would always be decremented down to 0 if the bit at
position 1 was set, meaning that N - 1 bits would be serialized instead
of N bits.

To fix this, the last_pos variable was introduced to track the actual
last position in value that is set.