de: Fix compile errors for untagged unions with multiple variants
If an untagged union has multiple variants, the original deserialization
logic would raise compile errors b/c the unrolled loops going over each
variant and attempting to deserialize into each one would each contain a
call to payloadSeed, which would return the variant type corresponding
to that loop. This means we had N return statements in
deserializeUntaggedUnion, each with a different return type.

To fix this, I added a check in payloadSeed to return an error instead
of a payload if the seed's type (i.e., the variant we're currently
deserializing into) isn't the same as the type of the payload we already
successfully deserialized.