de: Use std.math.cast in integer visitor
Thanks to fredi for finding this!

With @intCast, if the incoming integer does not fit into the destination
type, UB could result. This placed the onus on deserializers to ensure
that the integer they passed to the visitor is always correct.

Instead, we now use std.math.cast to convert between integers types. An
overflow error is returned if a type cannot be correctly cast. This
should be more safe and ease the work of deserializers.